[REFACTOR] Componentização por composição em componente do pacote collections
