+++ components_smells.csv
@@ -1,0 +1,27 @@
+id,file,Component,Smell,Details,"Em uma escala de 1 a 5, como você avalia a relevância desse smell?
+1 = smell não é importante e, provavelmente, não será refatorado
+5 = smell é muito importante e, provavelmente, será refatorado em breve.",Observação
+1,C:\Program Files\projects\cheathub\frontend\src\components\collections\collection\index.tsx,CollectionItem,Large Component,LOC: 142; Number of props: 18; Number of attributes: 0; Number of methods: 0,,
+2,C:\Program Files\projects\cheathub\frontend\src\components\collections\collection\index.tsx,CollectionItem,Too many props,"Number of props: 18; Props: collection,id,index,setId,expanded,setExpanded,setSelections,selectedSnippetId,setSelectedSnippetId,setExpandedSnippet,setHeading,readonly,bg,open,collapsed,duration,ease,borderRadius",,
+3,C:\Program Files\projects\cheathub\frontend\src\components\collections\crud\index.tsx,CollectionCrud,Large Component,LOC: 246; Number of props: 21; Number of attributes: 0; Number of methods: 0,,
+4,C:\Program Files\projects\cheathub\frontend\src\components\collections\crud\index.tsx,CollectionCrud,Too many props,"Number of props: 21; Props: name,setName,handleSubmit,handleDelete,snippets,setSnippets,handleCancel,setAlert,editing,deleting,submitting,snippetsOptions,open,collapsed,duration,ease,menu,borderRadius,borderColor,spacing,colors",,
+5,C:\Program Files\projects\cheathub\frontend\src\components\collections\snippets\index.tsx,SnippetItem,Large Component,LOC: 160; Number of props: 17; Number of attributes: 0; Number of methods: 0,,
+6,C:\Program Files\projects\cheathub\frontend\src\components\collections\snippets\index.tsx,SnippetItem,Too many props,"Number of props: 17; Props: snippet,index,selectedSnippetId,setSelectedSnippetId,expandedSnippet,setExpandedSnippet,editing,username,setTags,faveSnippet,handleFave,faving,border,open,collapsed,duration,ease",,
+7,C:\Program Files\projects\cheathub\frontend\src\components\navigation\actions\index.tsx,Actions,Large Component,LOC: 206; Number of props: 13; Number of attributes: 0; Number of methods: 0,,
+8,C:\Program Files\projects\cheathub\frontend\src\components\search\search-bar.tsx,SearchBar,Large Component,LOC: 137; Number of props: 18; Number of attributes: 0; Number of methods: 0,,
+9,C:\Program Files\projects\cheathub\frontend\src\components\search\search-bar.tsx,SearchBar,Too many props,"Number of props: 18; Props: searchText,onSearchTextChange,language,onLanguageChange,tags,onTagsChange,allTags,resetAll,heading,children,value,label,base,lg,open,collapsed,duration,ease",,
+10,C:\Program Files\projects\cheathub\frontend\src\containers\default.container.tsx,Page,Large Component,LOC: 120; Number of props: 22; Number of attributes: 0; Number of methods: 0,,
+11,C:\Program Files\projects\cheathub\frontend\src\containers\default.container.tsx,Page,Too many props,"Number of props: 22; Props: primary,primaryHeading,primaryChildren,primaryFooterHeading,primaryFooterSubheading,secondary,secondaryHeading,secondaryChildren,secondaryFooterHeading,secondaryFooterSubheading,preSecondaryChildren,postSecondaryChildren,modals,drawers,pagination,search,icon,width,background,borderRadius,base,lg",,
+12,C:\Program Files\projects\cheathub\frontend\src\context\appdata.context.tsx,AppDataProvider,Large Component,LOC: 135; Number of props: 24; Number of attributes: 0; Number of methods: 6,,
+13,C:\Program Files\projects\cheathub\frontend\src\context\appdata.context.tsx,AppDataProvider,Too many props,"Number of props: 24; Props: children,label,value,target,loadInitialData,loadResultsData,loadAllTags,data,setData,searchText,setSearchText,onSearchTextChange,language,setLanguage,onLanguageChange,tags,setTags,onTagChange,page,setPage,loading,setLoading,allTags,setAllTags",,
+14,C:\Program Files\projects\cheathub\frontend\src\context\collectionhandler.tsx,CollectionHandlerProvider,Large Component,LOC: 244; Number of props: 35; Number of attributes: 0; Number of methods: 4,,
+15,C:\Program Files\projects\cheathub\frontend\src\context\collectionhandler.tsx,CollectionHandlerProvider,Too many props,"Number of props: 35; Props: children,accessToken,loadCollectionsData,loadSnippetsOptions,label,value,duration,isClosable,render,selected,setSelected,selections,setSelections,loadingSelected,setLoadingSelected,selectedId,setSelectedId,handleSubmit,handleDelete,handleCancel,id,setId,name,setName,snippets,setSnippets,submitting,setSubmitting,deleting,setDeleting,editing,setEditing,alert,setAlert,clearValues",,
+16,C:\Program Files\projects\cheathub\frontend\src\context\datahandler.context.tsx,DataHandlerProvider,Large Component,LOC: 315; Number of props: 46; Number of attributes: 0; Number of methods: 5,,
+17,C:\Program Files\projects\cheathub\frontend\src\context\datahandler.context.tsx,DataHandlerProvider,Too many props,"Number of props: 46; Props: children,username,accessToken,loadInitialData,loadSnippetsData,loadFaveSnippets,loadCollectionsOptions,loadCollectionsData,loadSnippetsOptions,duration,isClosable,render,title,setTitle,value,setValue,description,setDescription,language,setLanguage,tags,setTags,source,setSource,privatize,setPrivatize,id,setId,submitting,setSubmitting,deleting,setDeleting,editing,setEditing,alert,setAlert,heading,setHeading,clearValues,faving,faveSnippet,setFaveSnippet,handleFave,handleDelete,handleCancel,handleSubmit",,
+18,C:\Program Files\projects\cheathub\frontend\src\context\profiledata.context.tsx,ProfileDataProvider,Large Component,LOC: 121; Number of props: 26; Number of attributes: 0; Number of methods: 5,,
+19,C:\Program Files\projects\cheathub\frontend\src\context\profiledata.context.tsx,ProfileDataProvider,Too many props,"Number of props: 26; Props: children,username,loggedIn,label,value,loadSnippetsData,loadCollectionsData,collectionsProfile,setCollectionsProfile,loadingCollections,setLoadingCollections,snippetsProfile,setSnippetsProfile,loadingSnippets,setLoadingSnippets,loadFaveSnippets,setFaveSnippets,faveSnippets,loadingFaves,setLoadingFaves,snippetsOptions,setSnippetsOptions,loadSnippetsOptions,collectionsOptions,setCollectionsOptions,loadCollectionsOptions",,
+20,C:\Program Files\projects\cheathub\frontend\src\pages\about.tsx,ProfileCard,Large Component,LOC: 135; Number of props: 8; Number of attributes: 0; Number of methods: 0,,
+21,C:\Program Files\projects\cheathub\frontend\src\pages\home.tsx,Home,Large Component,LOC: 89; Number of props: 16; Number of attributes: 0; Number of methods: 0,,
+22,C:\Program Files\projects\cheathub\frontend\src\pages\home.tsx,Home,Too many props,"Number of props: 16; Props: match,username,setUsername,email,setEmail,password,setPassword,returning,setReturning,loading,handleSignIn,data,setTags,handleFave,faveSnippet,faving",,
+23,C:\Program Files\projects\cheathub\frontend\src\pages\snippets.tsx,Snippets,Large Component,LOC: 166; Number of props: 20; Number of attributes: 0; Number of methods: 1,,
+24,C:\Program Files\projects\cheathub\frontend\src\pages\snippets.tsx,Snippets,Too many props,"Number of props: 20; Props: data,loading,searchText,setSearchText,onSearchTextChange,language,setLanguage,onLanguageChange,tags,setTags,onTagChange,page,setPage,allTags,loadInitialData,faveSnippet,handleFave,faving,username,accessToken",,
