+++ frontend/frontend.und/settings.xml
@@ -1,0 +1,263 @@
+<project>
+  <files
+    default_encoding="System"
+    readonly="false"
+    maxfilesize="10485760"
+    enforce_portability="false">
+    <dir
+      name=".">
+      <watched
+        excludes=".*"
+        subdirs="true">
+        <filter
+          name="CSS"/>
+        <filter
+          name="Html"/>
+        <filter
+          name="Javascript"/>
+        <filter
+          name="Php"/>
+        <filter
+          name="TypeScript"/>
+        <filter
+          name="Xml"/>
+      </watched>
+      <dir
+        name="public"/>
+      <dir
+        name="src">
+        <dir
+          name="components">
+          <dir
+            name="collections">
+            <dir
+              name="collection"/>
+            <dir
+              name="crud"/>
+            <dir
+              name="snippets"/>
+            <dir
+              name="ui"/>
+          </dir>
+          <dir
+            name="layout">
+            <dir
+              name="main">
+              <dir
+                name="areas"/>
+              <dir
+                name="grid"/>
+            </dir>
+            <dir
+              name="root">
+              <dir
+                name="areas"/>
+              <dir
+                name="grid"/>
+            </dir>
+          </dir>
+          <dir
+            name="modals"/>
+          <dir
+            name="navigation">
+            <dir
+              name="account"/>
+            <dir
+              name="actions"/>
+            <dir
+              name="navbar"/>
+            <dir
+              name="pagination"/>
+          </dir>
+          <dir
+            name="registration"/>
+          <dir
+            name="search"/>
+          <dir
+            name="shared"/>
+          <dir
+            name="snippet">
+            <dir
+              name="body"/>
+            <dir
+              name="card"/>
+            <dir
+              name="crud"/>
+            <dir
+              name="editor"/>
+          </dir>
+        </dir>
+        <dir
+          name="constants"/>
+        <dir
+          name="containers"/>
+        <dir
+          name="context"/>
+        <dir
+          name="lib"/>
+        <dir
+          name="pages"/>
+        <dir
+          name="services"/>
+        <dir
+          name="styles"/>
+      </dir>
+    </dir>
+  </files>
+  <languages>
+    <language
+      name="Web"/>
+  </languages>
+  <file_types/>
+  <annotations_options/>
+  <codecheck_options/>
+  <comparison_options
+    comparison_db=""
+    git_repo=""
+    git_commit=""
+    diff_flags="1"/>
+  <dependency_options
+    exclude_standard="true"
+    use_includes="true"
+    mode="link"/>
+  <imports/>
+  <metrics
+    declared_in_file_display_mode="no_path"
+    file_display_mode="no_path"
+    show_declared_in_file="false"
+    show_function_parameter_types="false"
+    write_column_titles="true"
+    add_uniquename_column="false"
+    metric_filter="28532"/>
+  <reports
+    date_stamping="false"
+    display_parameters="false"
+    index_by_method="false"
+    group_by_directory="false"
+    filename_mode="short"
+    generation_status="needs_generation"
+    generation_time="0"
+    html_clear_directory="false"
+    html_generation="true"
+    html_mode="alpha"
+    headertext_mode="false"
+    headerusertext_mode="false"
+    headertext_text="Report generated by Understand"
+    headerusertext_text=""
+    html_size="250"
+    text_clear_directory="false"
+    text_generation="true"
+    text_mode="single">
+    <report
+      name="Data Dictionary"
+      enabled="true"/>
+    <report
+      name="File Contents"
+      enabled="true"/>
+    <report
+      name="Program Unit Cross Reference"
+      enabled="true"/>
+    <report
+      name="Object Cross Reference"
+      enabled="true"/>
+    <report
+      name="Type Cross Reference"
+      enabled="true"/>
+    <report
+      name="Macro Cross Reference"
+      enabled="true"/>
+    <report
+      name="Include File Cross Reference"
+      enabled="true"/>
+    <report
+      name="Declaration Tree"
+      enabled="true"/>
+    <report
+      name="Extend Tree"
+      enabled="false"/>
+    <report
+      name="Invocation Tree"
+      enabled="false"/>
+    <report
+      name="Simple Invocation Tree"
+      enabled="true"/>
+    <report
+      name="Imports"
+      enabled="true"/>
+    <report
+      name="With Tree"
+      enabled="true"/>
+    <report
+      name="Simple With Tree"
+      enabled="true"/>
+    <report
+      name="Generic Instantiation"
+      enabled="true"/>
+    <report
+      name="Exception Cross Reference"
+      enabled="true"/>
+    <report
+      name="Renames"
+      enabled="true"/>
+    <report
+      name="Program Unit Complexity"
+      enabled="true"/>
+    <report
+      name="Project Metrics"
+      enabled="true"/>
+    <report
+      name="Program Unit Metrics"
+      enabled="true"/>
+    <report
+      name="File Metrics"
+      enabled="true"/>
+    <report
+      name="File Average Metrics"
+      enabled="true"/>
+    <report
+      name="Fortran Extension Usage"
+      enabled="true"/>
+    <report
+      name="Class Metrics"
+      enabled="true"/>
+    <report
+      name="Class OO Metrics"
+      enabled="true"/>
+    <report
+      name="Implicitly Declared Objects"
+      enabled="true"/>
+    <report
+      name="Uninitialized Items"
+      enabled="true"/>
+    <report
+      name="Unused Objects and Functions"
+      enabled="true"/>
+    <report
+      name="Unused Objects"
+      enabled="true"/>
+    <report
+      name="Unused Types"
+      enabled="true"/>
+    <report
+      name="Unused Program Units"
+      enabled="true"/>
+    <report
+      name="Uses Not Needed"
+      enabled="true"/>
+    <report
+      name="Withs Not Needed"
+      enabled="true"/>
+  </reports>
+  <web_options
+    add_found_imported_js_files="true"
+    allow_asp_style_php_tags="false"
+    allow_jquery="false"
+    allow_node_js="false"
+    allow_short_php_tags="true"
+    js_module_search_path=""
+    node_js_predefined_config_file="$CONF:/understand/javascript/nodejs_predefined.txt"
+    node_js_search_path=""
+    save_php_comments="true"
+    search_strings_for_entity_names="false"
+    php_version="8.3"/>
+</project>
