+++ frontend/package.json
@@ -17,7 +17,6 @@
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.22",
     "@types/node": "^14.14.37",
-    "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-select": "^4.0.15",
@@ -71,6 +70,7 @@
     "styled-components": "^5"
   },
   "devDependencies": {
+    "@types/react": "^18.3.12",
     "@typescript-eslint/eslint-plugin": "^4.21.0",
     "@typescript-eslint/parser": "^4.21.0",
     "babel-eslint": "^10.1.0",
